data is moved to phrases.swift new random generator is being used
