+++ Shamaz/phrases.swift
@@ -1,0 +1,21 @@
+//
+//  phrases.swift
+//  Shamaz
+//
+//  Created by Alexandra Ivanova on 30/03/2020.
+//  Copyright © 2020 Alex Ivanova. All rights reserved.
+//
+
+import Foundation
+
+struct Phrases {
+    let futureStoryStartPhrases = ["What would you like to be doing", "What would you like to have for dinner", "Where would you like to be", "Who would you like to meet", "Who would you like to be with"]
+    
+    let pastStoryStartPhrases = ["What did you do", "Where were you", "What did you have for breakfast", "What did you have for dinner", "What did you wear", "Who did you talk to", "Who did you meet"]
+    
+    let timeIndicators = ["1 day", "2 days", "3 days", "4 days", "5 days", "1 week", "3 weeks", "1 month", "3 months", "6 months", "1 year", "5 years", "10 years"]
+    
+    let futureStoryEndPhrase = "from now"
+    
+    let pastStoryEndPhrase = "ago"
+}
